requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
anyio==4.11.0
    # via
    #   httpx
    #   jupyter-server
appnope==0.1.4
    # via ipykernel
argon2-cffi==25.1.0
    # via jupyter-server
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==3.0.0
    # via stack-data
async-lru==2.0.5
    # via jupyterlab
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
babel==2.17.0
    # via jupyterlab-server
beautifulsoup4==4.14.2
    # via
    #   -r requirements.in
    #   nbconvert
black==25.9.0
    # via -r requirements.in
bleach==6.2.0
    # via nbconvert
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via argon2-cffi-bindings
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via black
comm==0.2.3
    # via
    #   ipykernel
    #   ipywidgets
debugpy==1.8.17
    # via ipykernel
decorator==5.2.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
executing==2.2.1
    # via stack-data
fastjsonschema==2.21.2
    # via nbformat
fqdn==1.5.1
    # via jsonschema
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via jupyterlab
idna==3.10
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
ipykernel==6.30.1
    # via
    #   -r requirements.in
    #   jupyter
    #   jupyter-console
    #   jupyterlab
ipython==9.6.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-pygments-lexers==1.1.1
    # via ipython
ipywidgets==8.1.7
    # via jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
json5==0.12.1
    # via jupyterlab-server
jsonpointer==3.0.0
    # via jsonschema
jsonschema==4.25.1
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2025.9.1
    # via jsonschema
jupyter==1.1.1
    # via -r requirements.in
jupyter-client==8.6.3
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.12.0
    # via jupyter-server
jupyter-lsp==2.3.0
    # via jupyterlab
jupyter-server==2.17.0
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab==4.4.9
    # via
    #   jupyter
    #   notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.27.3
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.15
    # via ipywidgets
lark==1.3.0
    # via rfc3987-syntax
markupsafe==3.0.3
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mistune==3.1.4
    # via nbconvert
mypy-extensions==1.1.0
    # via black
nbclient==0.10.2
    # via nbconvert
nbconvert==7.16.6
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
notebook==7.4.7
    # via jupyter
notebook-shim==0.2.4
    # via
    #   jupyterlab
    #   notebook
numpy==2.3.3
    # via pandas
packaging==25.0
    # via
    #   black
    #   ipykernel
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
pandas==2.3.3
    # via -r requirements.in
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.5
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
platformdirs==4.4.0
    # via
    #   black
    #   jupyter-core
prometheus-client==0.23.1
    # via jupyter-server
prompt-toolkit==3.0.52
    # via
    #   ipython
    #   jupyter-console
psutil==7.1.0
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.3
    # via stack-data
pycparser==2.23
    # via cffi
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   jupyter-console
    #   nbconvert
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
    #   pandas
python-json-logger==3.3.0
    # via jupyter-events
pytokens==0.1.10
    # via black
pytz==2025.2
    # via
    #   -r requirements.in
    #   pandas
pyyaml==6.0.3
    # via jupyter-events
pyzmq==27.1.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.32.5
    # via jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rfc3987-syntax==1.1.0
    # via jsonschema
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
ruff==0.13.3
    # via -r requirements.in
send2trash==1.8.3
    # via jupyter-server
setuptools==80.9.0
    # via jupyterlab
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via anyio
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.43
    # via -r requirements.in
stack-data==0.6.3
    # via ipython
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.4.0
    # via bleach
tornado==6.5.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
traitlets==5.14.3
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
types-python-dateutil==2.9.0.20250822
    # via arrow
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   sqlalchemy
tzdata==2025.2
    # via pandas
uri-template==1.3.0
    # via jsonschema
urllib3==2.5.0
    # via requests
wcwidth==0.2.14
    # via prompt-toolkit
webcolors==24.11.1
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.8.0
    # via jupyter-server
widgetsnbextension==4.0.14
    # via ipywidgets
